+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Introduced `mobx.configure({ disableErrorBoundaries })`, for easier debugging of exceptoins. By [NaridaL](https://github.com/NaridaL) through [#1262](https://github.com/mobxjs/mobx/pull/1262)
 * Introduced `@computed(options)`
 * `autorun`, `reaction` and `when` now can take an `onError` property as parameter
+* `toJS` now accepts options: `{    detectCycles?: boolean    exportMapsAsObjects?: boolean }`, both `true` by default
 
 ## Breaking changes